feat: Only load auto links in viewer to prevent them from being hardcoded in editor
